tree construction moved from grouping stage to client side node manager
