+++ packages/ag-grid-enterprise/src/treeData/treeNodeManager/treeNode.ts
