+++ packages/ag-grid-enterprise/src/treeData/treeNodeManager/treeRow.ts
